Adapt to subvisual's workflow
Why:

The example application was too complex and badly structured,
to work as a starting point. It would have to be removed on the beginning
of every project.

Some are use derequire, so it was necessary to expose react as a global
variable.

For styles the postcss-loader is also a must have.

Overall there were also some inconsistencies on the coding style.